requirements.txt
appdirs==1.4.3
click==6.7
dominate==2.3.1
Flask==0.12.1
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.2
gitdb2==2.0.0
GitPython==2.1.3
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
MySQL-python==1.2.5
packaging==16.8
pyparsing==2.2.0
six==1.10.0
smmap2==2.0.1
SQLAlchemy==1.1.9
visitor==0.1.3
Werkzeug==0.12.1
